geo-testing/requirements.txt
playwright>=1.55.0,<2.0.0
pymongo>=4.15.3,<5.0.0
python-dotenv>=1.2.1,<2.0.0

